rename RatingDialog as FeedbackDialog, add dialog style from call theme
