+++ bandyer-android-design/src/main/java/com/bandyer/sdk_design/feedback/BaseRatingBar.kt
@@ -1,4 +1,4 @@
-package com.bandyer.sdk_design.rating
+package com.bandyer.sdk_design.feedback
 
 import android.content.Context
 import android.graphics.PorterDuff
