+++ bandyer-android-design/src/main/java/com/bandyer/sdk_design/feedback/RatingBar.kt
@@ -1,4 +1,4 @@
-package com.bandyer.sdk_design.rating
+package com.bandyer.sdk_design.feedback
 
 import androidx.annotation.FloatRange
 import androidx.annotation.IntRange
