--- bandyer-android-design/src/main/res/layout/bandyer_rating_dialog_layout.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:animateLayoutChanges="true"
-    tools:context="rating.RatingDialog">
-
-    <include
-        android:id="@+id/bandyer_rating_layout"
-        layout="@layout/fragment_rating_layout"/>
-
-    <include
-        android:id="@+id/bandyer_rating_done_layout"
-        android:visibility="gone"
-        layout="@layout/fragment_rating_done_layout"/>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
