+++ bandyer-android-design/src/main/res/layout/fragment_feedback_sent_layout.xml
