+++ bandyer-android-design/src/main/res/values/themes_call.xml
@@ -34,6 +34,8 @@
         <attr name="bandyer_screenShareDialogStyle" format="reference" />
         <attr name="bandyer_fileShareDialogStyle" format="reference" />
 
+        <attr name="bandyer_feedbackDialogStyle" format="reference" />
+
         <attr name="bandyer_smartGlassDialogMenuStyle" format="reference" />
     </declare-styleable>
 
@@ -52,6 +54,7 @@
         <item name="bandyer_whiteboardDialogStyle">@style/BandyerSDKDesign.BottomSheetDialog.Whiteboard</item>
         <item name="bandyer_screenShareDialogStyle">@style/BandyerSDKDesign.BottomSheetDialog.ScreenSharePicker</item>
         <item name="bandyer_fileShareDialogStyle">@style/BandyerSDKDesign.BottomSheetDialog.FileShare</item>
+        <item name="bandyer_feedbackDialogStyle">@style/BandyerSDKDesign.FragmentDialog.Feedback</item>
     </style>
 
     <style name="BandyerSDKDesign.Theme.Call.Day.V16" parent="BandyerSDKDesign.Theme.Day">
@@ -69,6 +72,7 @@
         <item name="bandyer_whiteboardDialogStyle">@style/BandyerSDKDesign.BottomSheetDialog.Whiteboard</item>
         <item name="bandyer_screenShareDialogStyle">@style/BandyerSDKDesign.BottomSheetDialog.ScreenSharePicker</item>
         <item name="bandyer_fileShareDialogStyle">@style/BandyerSDKDesign.BottomSheetDialog.FileShare</item>
+        <item name="bandyer_feedbackDialogStyle">@style/BandyerSDKDesign.FragmentDialog.Feedback</item>
     </style>
 
     <style name="BandyerSDKDesign.Theme.Call.Night.V16" parent="BandyerSDKDesign.Theme.Night">
@@ -86,6 +90,7 @@
         <item name="bandyer_whiteboardDialogStyle">@style/BandyerSDKDesign.BottomSheetDialog.Whiteboard</item>
         <item name="bandyer_screenShareDialogStyle">@style/BandyerSDKDesign.BottomSheetDialog.ScreenSharePicker</item>
         <item name="bandyer_fileShareDialogStyle">@style/BandyerSDKDesign.BottomSheetDialog.FileShare</item>
+        <item name="bandyer_feedbackDialogStyle">@style/BandyerSDKDesign.FragmentDialog.Feedback</item>
     </style>
 
     <style name="BandyerSDKDesign.Theme.Call.DayNight" parent="BandyerSDKDesign.Theme.Call.DayNight.V16"/>
