+++ demo_app/src/main/res/values/strings.xml
@@ -35,7 +35,7 @@
     <string name="smart_glass_menu">Smartglass menu</string>
     <string name="screen_share_picker">Screen share picker</string>
     <string name="file_share">File share</string>
-    <string name="rating">Rating</string>
+    <string name="feedback">Feedback</string>
     <string name="add_upload">Add upload</string>
     <string name="add_download">Add download</string>
     <string name="add_download_available">Add download available</string>
